move all util files to util folder
