+++ __tests__/systemCalls.spec.ts
@@ -2,7 +2,7 @@ import { Arrays, Base58, MockVM, StringBytes, System, Crypto, Base64 } from "../
 import * as chain from '../assembly/koinos-proto-as/koinos/chain/chain';
 import * as protocol from '../assembly/koinos-proto-as/koinos/protocol/protocol';
 import * as authority from '../assembly/koinos-proto-as/koinos/chain/authority';
-import { test as TestObject } from "./test";
+import * as TestObject from "./test";
 
 const mockAccount = Base58.decode('1DQzuCcTKacbs9GGScRTU1Hc8BsyARTPqe');
 const mockAccount2 = Base58.decode('1DQzuCcTKacbs9GGScRTU1Hc8BsyARTPqE');
