+++ assembly/index.ts
@@ -3,10 +3,7 @@ export { Protobuf } from 'as-proto';
 export { u128, u128Safe } from 'as-bignum';
 
 export { System } from './systemCalls';
-export { Token } from './token';
 export * from './util';
-export { MockVM } from './mockVM';
-export { SafeMath } from './safeMath';
 
 export * from './koinos-proto-as/google/protobuf/any';
