+++ assembly/util/index.ts
@@ -3,3 +3,6 @@ export { Base58 } from "./base58";
 export { StringBytes } from "./stringBytes";
 export { Crypto } from "./crypto";
 export { Arrays } from "./arrays";
+export { SafeMath } from "./safeMath";
+export { MockVM } from "./mockVM";
+export { Token } from "./token";
